Migrate invite_info tests
